+++ iam_instance_profile.tf
@@ -1,0 +1,5 @@
+resource "aws_iam_instance_profile" "profile" {
+  count = local.create_iam_instance_profile
+  name  = var.iam_instance_profile_name
+  role  = var.iam_role_name
+}